Fix unsupported operand error when using baggage
In Tracer, we initialize baggage with `null` for root span.
This causes the baggage to be passed onto SpanContext also as `null`
and will cause an exception when we try to merge old and new baggage
array with `+` operand since it will be `[] + null`.

The best fix is probably to ensure baggage is always an array in
SpanContext constructor. The param type hint and `$baggage` property
type hint both uses `array`, so it doesn't make much sense to accept
non-arrays.