Support configuring tags to be added to each span
Jaeger has a concept of tracer/process tags,[1] which is a list of
tags that should be added to each span started by a given tracer.
This functionality is already implemented in jaeger-client-php,
but currently, there is no configuration option to allow users
to actually define these tags.

This patch extends the Config class to accept a new optional key
'tags', which is a key-value array representing tags that should
be added to each span started by the tracer built from this config.
Some unit tests were added for the new functionality.

---
[1] https://www.jaegertracing.io/docs/1.17/client-features/