Migrate away from PHPUnit's expectedException annotations
PHPUnit provides expectedException and related annotations to
declaratively specify information about the exception a test
is expected to throw. This API has been deprecated in PHPUnit 8
and is due to be removed in PHPUnit 9.[1]

This patch migrates the tests to use the recommended expectException()
and related methods instead of the annotations to ensure that a future
PHPUnit upgrade will not break them.

---
[1] https://github.com/sebastianbergmann/phpunit/issues/3332