Allow to configure whether to use one span per RPC
Currently, tracers initialized from the Config object are always
configured to follow Zipkin's model of using one span per RPC,
as opposed to the model where a span is always exclusive to a given
tracer (i.e. separate spans on the client and server).

If upstream applications do not follow this model, this can cause
unwanted span ID duplication in traces. This patch makes the option
configurable so that the appropriate behavior can be selected depending
on one's setup. To ensure backwards compatibility, the tracer continues
to default to the Zipkin model if no configuration was set.